requirements.txt
streamlit
pdfplumber
spacy
ultralytics
scikit-learn
torch
torchvision
langchain
opencv-python-headless
numpy
Pillow